Update Playlist tests
There’s a broken "resolve the inserted playlist" test. We need to fix
the following error, but I’m not sure how:

1) PlaylistService when creating a playlist should resolve the inserted
playlist:
     AssertionError: expected { Object (dataValues,
_previousDataValues, ...) } to have a property 'items'

There’s also this error, which I am not sure if it’s related to the
previous one or not:

  2) PlaylistService when creating a playlist should have inserted a
playlist in the database:
     Error: timeout of 2000ms exceeded. Ensure the done() callback is
being called in this test.