Init User model and service
The goal of this model and service is to synchronize a user’s
subscribed podcasts and playlists between the mobile app and web app.

Also, I accidentally committed the User tests in the last “Style
changes” commit :…(

I forgot to remember to scroll down in the Github Desktop UI before
committing, I didn’t want to commit those.