Workaround for playback stopping when converting to condensed view
Previously I was using .insertBefore to move the player element into
the condensed element when switching into the condensed view. It turns
out this causes the <audio> element to stop playing.

If I knew that, I would have implemented this full/condensed player
feature totally differently, but this commit provides a workaround to
use as much of the existing HTML as possible while resolving the
playback stopping bug.